add self links to HTML
FOO will link to its own <LI> element in this:

    - [function] FOO

This makes it easier to find links to individual functions in the doc.

Also, wrap the locative type, the symbol and the arglist in <SPAN>
elements of class "locative-type", "reference-object", and
"locative-args", respectively. It is now the CSS file's responsibility
to typeset these.

Also, in addition to the self link of section headings provided by
*DOCUMENT-FANCY-HTML-NAVIGATION* always make the heading text itself a
self link.

style.css was updated, so copy it over if you included it in your
module.