only change *PACKAGE* and *READTABLE* for SECTIONs
... when generating documentation.

Previously, if there was no containing section *PACKAGE* was bound to
the keyword package and linking didn't work well with symbols without
explicit packages.

Note that in general it is impossible to determine the right *PACKAGE*
to be used for a docstring so using the current package at least gives
the caller some control.