stabilize and unbreak URLs for the common case
Previously, symbols were printed with PRIN1 and that formed a part of
their URL. The problem with that was that for example a non-uppercase
name would be escaped in an implementation specific
manner (http://www.lispworks.com/documentation/lw71/CLHS/Body/02_cd.htm).

So we turn symbols into URL fragments with a simple home-grown method,
which is equivalent to PRIN1 for exported symbols that print without
escapes. This should cover most cases, right?

See *DOCUMENT-URL-VERSIONS*.