change DOCUMENT's default arguments and return values
Previously, DOCUMENT's :STREAM was NIL by default, so output was
collected in a string. With this change, the default :STREAM is T,
with which output is sent to *STANDARD-OUTPUT*.

Also changed DOCUMENT's default :FORMAT from :MARKDOWN to :PLAIN.

Also, with PAGES NIL, DOCUMENT's return values are changed:

- If STREAM is NIL, then the output string is returned (previously, it
  was in a one element list).

- If STREAM is not NIL, then NIL is returned.

With these changes

    (document #'pax:locate)

prints the documentation of #'PAX:LOCATE to the standard output. The
documentation is stripped of markdown links and such, which makes this
a good default for interactive use in the repl

None of these changes are visible, if DOCUMENT is used via
PAX::@DOCUMENTATION-UTILITIES.