don't codify package and ASDF system names
It wasn't a good idea to codify anything that's not named by symbols.
Consider the docstring "URL". With this change, for it to be codified,
the symbol with name URL must be internal to the package current at
the time of documentation generation. This can be controlled by the
package author. On the other hand, previously "URL" would have been
codified also if someone defined a package or ASDF system named "URL",
which is not something the package author can reasonably control.