add interactive documentation browser for Slime
PAX documentation can be viewed live in Emacs with w3m. Autolinking
generates magic "pax:" URLs. Documentation is generated when such a
link is followed.

If src/pax.el is loaded into Emacs, the elisp function
mgl-pax-document can be used to generate documentation for a PAX URL.
The documentation is a single HTML page generated by PAX via Slime.
For example, to view the documentation of this very SECTION, one can
do:

    M-x mgl-pax-document
    View Documentation of: pax::@documentation-in-emacs

To generate an overview of the CL package, enter this instead:

    cl

To view the entire documentation of PAX:

    pax::@pax-manual

To bypass the disambiguation page for stuff with multiple definitions,
specify a PAX locative:

    null function