support arbitrary browsers for live documentation
See MGL-PAX::@BROWSING-LIVE-DOCUMENTATION.

When the value of the new Elisp variable `mgl-pax-browser-function` is
not `w3m-browse-url`, requests are served via a web server started in
the running Lisp, and documentation is most likely displayed in a
separate browser window .

By default, `mgl-pax-browser-function` is `nil`, which makes PAX use
`browse-url-browser-function`. You may want to customize the related
`browse-url-new-window-flag` or, for Chrome, set
`browse-url-chrome-arguments` to `("--new-window")`.

In the browser, clicking on the locative on the left of the
object (e.g. in `- [function] PRINT`) will raise and focus the Emacs
window (if Emacs is not in text mode, and also subject to window
manager focus stealing settings), then go to the corresponding source
location. For sections, clicking on the lambda link will do the
same (see *DOCUMENT-FANCY-HTML-NAVIGATION*).

Finally, note that the URLs exposed by the web server are subject to
change.