easier extensions for first-class objects
For example, when a SECTION was subclassed as CATEGORY, one had to
defin a CATEGORY locative, a CATEGORY-DREF, and all the LOCATE/RESOLVE
machinery to be able to specialize DOCUMENT-DREF or other generic
functions used to extend some operation.

With this change, this is no longer necessary because these generic
functions are now tried on the RESOLVEd definition (if possible), then
on the LOCATEd definition. The net effect is that when there is a
perfectly good first-class object for a definition, operations can be
extended by specializing on their type without needing to jump through
the locative/dref hoop.