resolve reflink title override ambiguity
Previously, when generating documentation for PAX,
[HUNCHENTOOT][package] was processed into a link to the PACKAGE
LOCATIVE because the HUNCHENTOOT package was not being documented, but
the PACKAGE LOCATIVE was.

With this change, the ambiguity in interperting [HUNCHENTOOT][package]
as a title override [1] or a specified locative [2] is resolved in
latter. Note that this ambiguity arises only when the Markdown
definition (here, "package") names a locative.

[1]: [HUNCHENTOOT][package] as title override: the link target
     is (DREF 'PACKAGE 'LOCATIVE), and the link title is HUNCHENTOOT.

[2]: [HUNCHENTOOT][package] as specified locative: the link target
     is (DREF "HUNCHENTOOT" 'PACKAGE), and the link title is the default for
     that DREF, HUNCHENTOOT.

See PAX::@UNSPECIFICED-LOCATIVE for the documentation.