add NOTEs
Add the macro NOTE and the locative type NOTE. Notes are like comments
with docstring and maybe a body. Named notes can gather the docstrings
of the other notes in their body and can be included with the
DOCSTRING locative.

Notes are intended to help reduce the distance between code and its
documentation when there is no convenient definition docstring to use
nearby.

This is an experimental feature.