document: include all PAX::@PUBLIC-SUPERCLASSES
The public superclasses of a class are tightest envelope of
superclasses with names exported from some package. This envelope is
constructed by recursing depth-first into the superclass hierarchy. If
the name of the superclass currently processed is exported from any
package, then it is collected as a public superclass, and we do
not recurse into its superclasses.